DELFT3DFM-1428: When a boundary condition does not have any data defined, the validation report will now show a validation error instead of a validation warning.
I had to add some extra functionality to tests in WaterFlowFMBoundaryConditionValidatorTest class, because boundary conditions did not contain data and were adding validation errors to the validation report.

I also renamed variables from "[name]Boundary" to "[name]BoundaryCondition" to make clear that the variable is a boundary condition and not a boundary.

Former-commit-id: c3423592095b553d195eaccefac3cd5dc5329aac
Former-commit-id: 0189fa0f673bb5afd054db9cc1eadf337c319542