Merged revision(s) 39343-39412 from trunk/delta-shell/Products/NGHS: D3DFMIQ-88: Removed the following projects: * DelftTools.OpenMI2 * DeltaShell.OpenMI2Wrapper * DeltaShell.OpenMIWrapper * DelftTools.OpenMI2.Tests * DeltaShell.OpenMI2Wrapper.Tests * DeltaShell.OpenMIWrapper.Tests * Deltares.IO.FewsPI * Deltares.IO.FewsPI.Test
Some relevant files from Deltares.IO.FewsPI and Deltares.IO.FewsPI.Test have been moved to the DeltaShell.Plugins.Fews and DeltaShell.Plugins.Fews.Tests projects. Other files have been removed from the whole solution.
........
D3DFMIQ-90: Added two tests for the toString method of the MappingColumn class to improve code coverage.
........
D3DFMIQ-90: Added tests that make sure that null is returned when importing with GroupableCloudImporter with a target that is null or not a list object. This makes the code coverage higher.
........
D3DFMIQ-89: Deleted the Waq-project from the setup solution. Some files have not been deleted, because they are used in other projects.
........
SOBEK3-1376 Handled exception with a better message. Created tests covering this handling. 
........
D3DFMIQ-125: Added DIMR technical reference manual to the welcome pages of all installers.
........
D3DFMIQ-86
Update product names of Delft3DFM and D-Hydro suites to 2018.01
........
D3DFMIQ-86
Update manuals in the installers for D-Hydro 1.4.3 release
........
D3DFMIQ-90: Created tests for adding, inserting, removing, replacing and clearing features in HydroNetworkFeatureCollection objects. These tests check mostly if the objects in the network within such an object is synchronized with the HydroNetworkFeatureCollection.
........
D3DFMIQ-63: The initial conditions file 'B8_initials.inc' now writes its INITIALS & DEFAULTS values on separate lines instead of on just one line. This is because Delwaq has a limit of 1000 characters. With a large amount of Initial Conditions, this would lead to a loss of data, with the old way of writing this data.
........
SOBEK3-1376 Removed useless comment.
........
D3DFMIQ-138
Split FM installer build targets, such that FM_Open and FM_RWS can now be built manually during the release only
Should significantly speed-up the build of FM_HM and FM_HMWQ on TeamCity
........
Improved message when killing delwaq exe (and made it warning level)
Also added sleep of 100 after killing the process to make it more stable
........
Made all tests use testoutput directory in bin folder for storing temporary test files (preparation for NUnit 3)
........
D3DFMIQ-109
Correcting spelling mistake 'Retrieving'
........
DELFT3DFM-1374: Improved the importing of dry points from large xyz-files by factor 7,5.
........
DELFT3DFM-1374: When giving group names to features when importing xyz files, the user sees the progress of this process.
........
DELFT3DFM-1374: Code review improvements. Mainly added comments to explain functionality. Also, I have increased the amount of time for a performance test I made earlier for this issue.
........
DELFT3DFM-1374: Fixed a failing test in HydroAreaGuiIntegrationTest. The importer that is used in this test was moved from the NetworkEditorApplicationPlugin to the FlowFMApplicationPlugin, so I added the FlowFMApplicationPlugin to the test. Also, re-ordered the logic in GroupablePointCloudImporter such that when a BaseFolder is not found (which means that no FM-model is present), the import will still go on.
........
DELFT3DFM-1374: Revert previous commit.
........
DELFT3DFM-1374: Fixed a failing test in HydroAreaGuiIntegrationTest. The importer that is used in this test was moved from the NetworkEditorApplicationPlugin to the FlowFMApplicationPlugin, so I added the FlowFMApplicationPlugin to the test. Also, re-ordered the logic in GroupablePointCloudImporter such that when a BaseFolder is not found (which means that no FM-model is present), the import will still go on. I moved this test to WaterFlowFMModelGuiIntegrationTest test class.
........
D3DFMIQ-136
Acceptance test for the FM models added

........
DELFT3DFM-896: When clearing the model output of a Water Quality Model, the right output data items are cleared or removed. Also, the right files are removed from the project as well.
........
DELFT3DFM-896: Updated a test. Instead of loading a project, the test is building the model up programmatically.
........
DELFT3DFM-1152: Added a field for setting the crest width to the weir settings view.
........
DELFT3DFM-896: Code review comments: when checking equality of two string collections, we now first check the count equality of both collections before checking the actual values.
........
DELFT3DFM-896: Added an extra check for the presence of an explicit working directory in case of deleting files at a WAQ-model output cleanup.
........
DELFT3DFM-1026
Added support for reading zValues from NetCdf files (UGrid only)
Added tests for existing UnstructuredGridFileHelper functionality
Existing SpatialOperations are now re-applied when switching between BedLevel locations
........
DELFT3DFM-1428: When a boundary condition does not have any data defined, the validation report will now show a validation error instead of a validation warning. 

I had to add some extra functionality to tests in WaterFlowFMBoundaryConditionValidatorTest class, because boundary conditions did not contain data and were adding validation errors to the validation report.

I also renamed variables from "[name]Boundary" to "[name]BoundaryCondition" to make clear that the variable is a boundary condition and not a boundary.
........
DELFT3DFM-1428: Removed a test that is testing DIMR functionality, not DeltaShell functionality. This test was failing because it tries to run a model that does not pass validation anymore due to changes to the validation report in this issue. 

Fixing the boundary conditions in the model will make the model pass validation and enables the model to run through DIMR. In the end, we do not want a model to validate successfully when it fails to run in DIMR. That is why we remove this test.
........
DELFT3DFM-1428: Changed an occasion of Assert.That to Assert.IsTrue to be more consistent throughout the whole code base.
........
DELFT3DFM-1428: Removed a test that is testing DIMR functionality, not DeltaShell functionality. This test was failing because it tries to run a model that does not pass validation anymore due to changes to the validation report in this issue. 

Fixing the boundary conditions in the model will make the model pass validation and enables the model to run through DIMR. In the end, we do not want a model to validate successfully when it fails to run in DIMR. That is why we remove this test.
........
DELFT3DFM-1428: Removed redundant using-statements from NHibernateHydroModelIntegrationTest class.
........
DELFT3DFM-1026
Process code review comments
........

Former-commit-id: 978e49cd098904423338491db6f6a3103ae7a631
Former-commit-id: cbcc40378c4172c02ce120b523903768f0b98979