FM1D2D-1936: Fix side view (#308)
* FM1D2D-1936: Fix side view

The side view throws InvalidOperationException when the computed chainages are not monotonous. This could happen for 2 distinct reasons:
1) the structure chainages are not ordered;
2) the shift which is applied before or after a structure's chainage is larger than the distance to the neighboring chainage;

Both situations have been addressed and unit tests have been added.

The SideViewFunctionCreator now assures that active structure chainages are ordered by increasing chainage.

Some refactoring has been done to improve the separation of responsibilities:
1) the internal class is now a 'standalone' calculator of the output data, and has complete control over allocating and populating the 'updated' arrays;
2) The FunctionUpdater class instantiates the calculator, assigns data retrieved from the IFunction, invokes the calculation and assigns the updated data to the IFunction.

* review rework: extracted calculator class

Renamed the calculator class, extracted it to separate file, extracted
applicable unit tests to separate file.

Some other minor refactoring such as renamed output properties of the
calculator, updated some comments.

* review rework, interface change and added tests

Fixed comments and added ensure for null arguments.

I changed the interface so that the Calculate receives all input data,
instead of the constructor. This way the class could be instantiated once
and the calculation be invoked multiple times.

Added a test for when the location chainage coincides with a structure
chainage. In such cases the location chainage is inserted first. This
consistently works out ok also if the chainage is at 0.0. The
implementation of the calculation was adjusted to fix these situations.

* review rework

* removed private properties,
* renamed updateIndex to outputIndex,
* structure shift is now a constant and renamed to offset,
* constructor removed because nothing happens in it anymore.

Former-commit-id: b7e6fc53c190b2df9f5505bec915380b215bdd7a