FM1D2D-2172: Update acceptance model expectations (#555)
After re-enable Pillar Bridge option is implemented/exposed in the code we now read in the SOBEK2 pillar bridges. We need to update the expectations of the acceptance model criteria. We now read in for the hunze models 1 bridge and create 1 compound structure in which we place the pillar bridge in. We also need to update the reference files in the avn acceptance test bench.

Former-commit-id: f6d9b7668fe2bea4f9ba1ce5dff9c82d7d3b7a4f